requirements.txt
pytest==7.1.2
Faker==9.8.3
jsonpath==0.82
redis==4.3.4
DingtalkChatbot==1.5.3
urllib3==1.26.7
PyMySQL==1.0.2
pyDes==2.0.1
requests_toolbelt==0.9.1
jinja2==3.1.2
pyyaml==6.0
mitmproxy==8.0.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6